+++ Wrappers/Python/test/test_dataexample.py
@@ -29,7 +29,7 @@ import numpy as np
 
 class TestTestData(CCPiTestClass):
     def test_noise_gaussian(self):
-        camera = dataexample.CAMERA.get()
+        camera = dataexample.CAMERA.get(dtype=np.float32)
         noisy_camera = noise.gaussian(camera, seed=1)
         norm = (camera - noisy_camera).norm()
         decimal = 4
